backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
asyncpg==0.29.0
pdfplumber==0.11.1
python-docx==1.1.2
anthropic==0.28.0
httpx==0.27.0
resend==2.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
